requirements.txt
flask
flask_sock
pillow
pywin32